release point indicator added. visual indicator that displays: relside, relheight, extension. Fixed the start up bug that does not allow any pitch modifications before playing the defualt pitch. This also allows for any pitch preset to be selected, and then pitched using the pitch button on the UI. added simple menu for the pitch stats. TODO: spawn implementation. Script optmization fixing DRY violation.
