+++ Assets/InfoMenu.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class pitchController : MonoBehaviour
+public class InfoMenu : MonoBehaviour
 {
     // Start is called before the first frame update
     void Start()
