+++ Assets/InfoMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 335d77dca284cf14da747cbfce065685
+guid: bcd3611ee05c8bc40a5778eccab0fc3d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
